chat_with_X_tutorials/chat_with_youtube_videos/requirements.txt
streamlit
asyncio
youtube-transcript-api==0.6.3
numpy
sentence-transformers
google-generativeai